% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMins.R
\name{colMins}
\alias{colMins}
\title{Form Column Minima}
\usage{
colMins(x, na.rm = FALSE)
}
\arguments{
\item{x}{data.frame or array of two dimensions, containing numeric, complex, integer or logical values.}

\item{na.rm}{logical: a logical indicating whether missing values should be removed.}
}
\value{
A numeric or complex data.frame/vector. For min and max if there are only numeric inputs and all are empty (after possible removal of NAs), the result is double (Inf or -Inf), consistent with min and max.
}
\description{
Calculates the minimum of each column of a matrix, a data.frame, or any 
other 2-dimensional structure.
}
\examples{
 abc = array(1:30, dim = c(6,5))
 colMins(abc)
 # returns: 1  7 13 19 25
 
 def = as.data.frame(abc)
 colMins(def)
 # returns
 #   V1 V2 V3 V4 V5 
 #    1  7 13 19 25 
}
\seealso{
colMaxs, rowMaxs, rowMins, colMeans, rowMeans, colSums, rowSums
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
