% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_time_excel2POSIXct.R
\name{conv_time_excel2POSIXct}
\alias{conv_time_excel2POSIXct}
\title{Convert MS Excel time stamp to POSIXct Format}
\usage{
conv_time_excel2POSIXct(time,
  reftime = as.POSIXct("1900-01-01 00:00 UTC"))
}
\arguments{
\item{time}{numeric; a time stamp in hours since 01/01/1900}

\item{reftime}{POSIXct, optional; provide another reference time}
}
\value{
time_POSIXct POSIXct: time in format POSIXct
}
\description{
Convert MS Excel time stamp to POSIXct Format.
}
\examples{

  # 31069 hours since 01/01/1900 00:00:00 => 22/01/1985 
  time_excel <- 31069 
  time_posixct <- conv_time_excel2POSIXct(time_excel)
  print(time_posixct)
}
