% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_time_POSIXct2excel.R
\name{conv_time_POSIXct2excel}
\alias{conv_time_POSIXct2excel}
\title{Convert POSIXct Format to MS Excel time stamp}
\usage{
conv_time_POSIXct2excel(time,
  reftime = as.POSIXct("1900-01-01 00:00 UTC"))
}
\arguments{
\item{time}{POSIXct; time in format POSIXct}

\item{reftime}{POSIXct, optional; provide another reference time}
}
\value{
time numeric; time as MS Excel time stamp in hours since 01/01/1900
}
\description{
Convert POSIXct Format to MS Excel time stamp
}
\examples{

  # 2001-03-29 23:00:00 UTC => 36980 hours since 1900-01-01 00:00 UTC
  time_posixct <- as.POSIXct('2001-03-29 23:00:00 UTC', tz='UTC')
  time_excel <- conv_time_POSIXct2excel(time_posixct)
  print(time_excel)
}
