% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_time_str2excel.R
\name{conv_time_str2excel}
\alias{conv_time_str2excel}
\title{Convert string/character time stamp to MS Excel time stamp}
\usage{
conv_time_str2excel(time, format = "\%Y-\%m-\%d \%H:\%M",
  reftime = "1900-01-01 00:00 UTC", refformat = "\%Y-\%m-\%d \%H:\%M")
}
\arguments{
\item{time}{character; time stamp as character}

\item{reftime}{POSIXct, optional; provide another reference time}
}
\value{
time numeric; time as MS Excel time stamp in hours since 01/01/1900
}
\description{
Convert string/character time stamp to MS Excel time stamp
}
\examples{

  # 2001-03-30 01:00:00 UTC => 36980 hours since 1900-01-01 00:00 UTC
  time_char <- '2001-03-30 01:00:00 UTC'
  time_excel <- conv_time_POSIXct2excel(time_char)
  print(time_excel)
}
