% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnb.R
\name{mnb}
\alias{mnb}
\title{Mean Normalized Bias}
\usage{
mnb(p, o, na.rm = TRUE)
}
\arguments{
\item{p}{numeric array; predicted values (model)}

\item{o}{numeric array; measured/observed values}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
numeric; mean normalized bias
}
\description{
Calculate a mean normalized bias.
}
\details{
MNB = 1/n * sum((p-o)/o)
}
\examples{

  # ex 1
  p <- 1:5
  o <- 2:6
  mnb(p,o)
  > -0.29
  
  # ex 2
  p <- c(1:4, NA)
  o <- 2:6
  mnb(p,o)
  > -0.3208333
  
  # ex 3
  p <- c(1:4,NA)
  o <- 2:6
  mnb(p,o,na.rm=FALSE)
  > NA
  
  # ex 4
  p <- 1:5
  o <- c(2:5, 0)
  mnb(p,o)
  > -0.3208333
  
  # ex 5
  p <- 1:5
  o <- c(2:5, 0)
  mnb(p,o, na.rm=FALSE)
  > Inf

}
