% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDaysOfMonth.R
\name{nDaysOfMonth}
\alias{nDaysOfMonth}
\title{Returns the number of days per month}
\usage{
nDaysOfMonth(month, year = 2001)
}
\arguments{
\item{month}{numeric or array of numerics between 1 and 12}

\item{year}{numeric, optional}
}
\value{
numeric or array of numerics giving the number of days per month
}
\description{
Returns the number of days per month
}
