% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDaysOfYear.R
\name{nDaysOfYear}
\alias{nDaysOfYear}
\title{Get number of days of provided year}
\usage{
nDaysOfYear(year)
}
\arguments{
\item{year}{numeric; a year}
}
\value{
numeric: numer of days of the particular year
}
\description{
Return the number of days of the year provided as input argument
}
\examples{

# a non-leap year:
nDaysOfYear(2001)
> 365

# a leap year:
nDaysOfYear(2004)
> 366
}
