% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{pmean}
\alias{pmean}
\title{Parallel mean}
\usage{
pmean(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric arguments (see Note).}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A vector of length the longest of the input vectors, or length zero if one of the inputs had zero length. The arithmetic mean of the values in '...' is computed, as a numeric vector. If '...' are not logical (coerced to numeric) or numeric (including integer), NA_real_ is returned, with a warning.
}
\description{
Returns the parallel mean of the input values.
 
 pmean takes one or more vectors as arguments, recycles them to common length and returns a single vector giving the ‘parallel’ mean of the argument vectors.
}
\examples{
  abc = 1:5
  def = 5:1
  pmean(abc, def)
  # returns: 3 3 3 3 3
  
  ghi = c(6, 6, 6, NA, 6)
  pmean(abc, def, ghi)
  # returns: 4 4 4 NA 4
  
  pmean(abc, def, ghi, na.rm = TRUE)
  # returns: 4 4 4 3 4
  
}
\seealso{
pmax, pmin
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
