% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.two.R
\name{pmean.two}
\alias{pmean.two}
\title{Parallel mean for two 1D or 2D numeric arrays}
\usage{
pmean.two(a1, a2, na.rm = FALSE)
}
\arguments{
\item{a1}{numeric array; 1D or 2D array with same number of rows as a2; if 1D, the length is considered as number of rows}

\item{a2}{numeric array; 1D or 2D array with same number of rows as a1; if 1D, the length is considered as number of rows}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
numeric array; length==nrow of a1 and a2 if one of them was 2D array; length=length of a1 and a2 if both were 1D arrays
}
\description{
Calculates parallel mean of two 1D or 2D numeric arrays. Other classes or sizes are probably not possible. Wrong inputs are not caught due to reasons of performance.
}
\examples{

# ex 1
b1 <- array(1:12, dim = c(4,3))
b2 <- 1:4
pmean.two(b1, b2)
> 4 5 6 7

# ex 2
b1 <- array(1:12, dim = c(4,3))
b2 <- array(1:12, dim = c(4,3))
pmean.two(b1, b2)
> 5 6 7 8

}
