% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMaxs.R
\name{rowMaxs}
\alias{rowMaxs}
\title{Form Row Maxima}
\usage{
rowMaxs(x, na.rm = FALSE)
}
\arguments{
\item{x}{data.frame or array of two dimensions, containing numeric, complex, integer or logical values.}

\item{na.rm}{logical: a logical indicating whether missing values should be removed.}
}
\value{
A numeric or complex data.frame/vector. For min and max if there are only numeric inputs and all are empty (after possible removal of NAs), the result is double (Inf or -Inf), consistent with min and max.
}
\description{
Calculates the maximum of each row of a matrix, a data.frame, or any 
other 2-dimensional structure.
}
\examples{
 abc = array(1:30, dim = c(6,5))
 rowMaxs(abc)
 # returns: 25 26 27 28 29 30
 
 def = as.data.frame(abc)
 rowMaxs(def)
 # returns
 #   25 26 27 28 29 30
}
\seealso{
colMaxs, colMins, rowMins, colMeans, rowMeans, colSums, rowSums
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
