% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max4plot.R
\name{get.max4plot}
\alias{get.max4plot}
\title{TODO}
\usage{
get.max4plot(x, type, x.max.default = 1)
}
\arguments{
\item{x}{numeric: one or more values for}

\item{type}{integer: type of}
}
\description{
Calculates the axis limits (min,max) for plotting a data series. Different 
types of methods can be choosen, such no rounding of min and max values and
rounding to the n_th digit.
}
\author{
Daniel Neumann, daniel.neumann@posteo.de
}
