% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDaysOfMonth.R
\name{nDaysOfMonth}
\alias{nDaysOfMonth}
\title{Returns the number of days per month}
\usage{
nDaysOfMonth(month, year = 2001)
}
\arguments{
\item{month}{numeric or array of numerics between 1 and 12}

\item{year}{numeric, optional [default: 2001; no leap year]}
}
\value{
numeric or array of numerics giving the number of days per month
}
\description{
Returns the number of days per month.
}
\examples{

# #days in may 2001:
nDaysOfMonth(5)
> 31

# #days in February 2001:
nDaysOfMonth(2)
> 28

# #days in February 2004:
nDaysOfMonth(2, 2004)
> 29

# #days in a set of months in 2007
nDaysOfMonth(c(1,2,6,12), 2007)
> 31 28 30 31
}
